requirements.txt
python-telegram-bot==20.3
python-dateutil==2.8.2
ics==0.7.2
google-generativeai
python-dotenv
timezonefinder
